refactor: clang fixes & .scripts cleanup
