+++ .scripts/configure.js
@@ -1,132 +1,18 @@
+const colors = require('colors');
 const {Command} = require('commander');
-const {twig} = require('twig');
 const fs = require('fs');
 const path = require('path');
 
-console.log(process.env.NODE_HEADERS);
+const {
+  packageJson,
+  supportedIdes,
+  supportedBuildSystems,
 
-const packageJson = JSON.parse(fs.readFileSync(path.join(__dirname, '..', 'package.json')).toString());
-
-/**
- * Colors for console.log
- */
-const colors = require('colors');
-
-/**
- * Supported IDE list
- */
-const supportedIdes = {
-  CLION: 'clion',
-  VSCODE: 'vscode',
-};
-
-/**
- * Supported Build Systems
- */
-const supportedBuildSystems = {
-  // https://cmake.org/ https://www.npmjs.com/package/cmake-js
-  CMAKE: 'cmake',
-  // https://gyp.gsrc.io/ https://github.com/nodejs/node-gyp
-  GYP: 'gyp',
-  // https://xmake.io/
-  XMAKE: 'xmake',
-};
-
-/**
- * node version
- */
-const nodeVersion = process.version.substr(1);
-
-/**
- * From process.platform
- */
-const nodeLibLocationGyp = {
-  aix: '',
-  android: '',
-  darwin: '',
-  freebsd: '',
-  linux: path.join(process.env.HOME || '~', '.cache', 'node-gyp', nodeVersion, 'include', 'node'),
-  openbsd: '',
-  sunos: '',
-  win32: path.join(process.env.TEMP || '', '..', 'node-gyp', 'Cache', nodeVersion, 'include', 'node'),
-};
-
-const nodeLibLocationCmake = {
-  aix: '',
-  android: '',
-  darwin: '',
-  freebsd: '',
-  linux: path.join(process.env.HOME || '~', '.cache', 'node-gyp', nodeVersion, 'include', 'node'),
-  openbsd: '',
-  sunos: '',
-  win32: path.join(
-    `${process.env.HOMEDRIVE}${process.env.HOMEPATH}`,
-    '.cmake-js',
-    `node-${process.arch}`,
-    process.version,
-    'include',
-    'node',
-  ),
-  // C:\Users\dragosc\.cmake-js\node-x64\v12.20.2\include\node
-};
-
-/**
- * Generic libs
- */
-
-const pathIsValid = (item) => {
-  try {
-    fs.statSync(item);
-  } catch (e) {
-    return false;
-  }
-  return true;
-};
-
-const haveNanInPackagesJson = () =>
-  ['dependencies', 'devDependencies'].map((item) => packageJson[item] && packageJson[item].nan).filter((item) => item)
-    .length > 0;
-
-const haveNapiInPackagesJson = () =>
-  ['dependencies', 'devDependencies']
-    .map((item) => packageJson[item] && packageJson[item]['node-addon-api'])
-    .filter((item) => item).length > 0;
-
-const libraryFolders = () => {
-  let folders = ['', 'openssl']
-    .map((item) => [item, path.join(nodeLibLocationGyp[process.platform], item)])
-    .reduce((acc, item) => [].concat(acc, item), [])
-    .filter(pathIsValid);
-
-  if (options.buildSystem === supportedBuildSystems.CMAKE || options.ide === supportedIdes.CLION) {
-    folders = ['', 'openssl']
-      .map((item) => [item, path.join(nodeLibLocationCmake[process.platform], item)])
-      .reduce((acc, item) => [].concat(acc, item), [])
-      .filter(pathIsValid);
-  }
-
-  // will be dependent on node-gyp files
-  // if (options.buildSystem === supportedBuildSystems.XMAKE) {}
-
-  if (haveNanInPackagesJson()) {
-    folders.push(path.join(__dirname, '..', 'node_modules', 'nan'));
-  }
-
-  // will be dependent on node-gyp files
-  if (options.buildSystem === supportedBuildSystems.XMAKE) {
-    folders = folders.map((item) => item.replace(/\\/g, '/'));
-  }
-
-  return folders;
-};
-
-const twigCompile = (name, twigValues) => {
-  const filePath = path.join(__dirname, 'configure', `${name}.twig`);
-  console.log(`Reading ${filePath}`.gray);
-  return twig({
-    data: fs.readFileSync(filePath).toString(),
-  }).render(twigValues);
-};
+  haveNanInPackagesJson,
+  haveNapiInPackagesJson,
+  libraryFolders,
+  twigCompile,
+} = require('./configure/lib');
 
 /**
  * CMake
@@ -138,13 +24,13 @@ const configureCMakeListsTxt = () => {
     fs.writeFileSync(
       filePath,
       twigCompile('CMakeLists.txt', {
-        folders: libraryFolders(),
+        folders: libraryFolders(options),
         hasNapi: haveNapiInPackagesJson(),
         hasNan: haveNanInPackagesJson(),
       }),
     );
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -159,12 +45,12 @@ const configureXmakeLua = () => {
     fs.writeFileSync(
       filePath,
       twigCompile('xmake.lua', {
-        folders: libraryFolders(),
+        folders: libraryFolders(options),
         platform: process.platform,
       }),
     );
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -185,6 +71,10 @@ const configureCompilerClear = () => {
     /build:addon:[\s]+/,
     'build:addon:gyp',
   );
+
+  packageJson.scripts['build:dev'] = 'npm run build:addon -- --debug';
+  packageJson.scripts['prebuild:dev'] = '';
+  delete packageJson.scripts['prebuild:dev'];
 };
 
 const configureCompilerCmake = () => {
@@ -216,6 +106,9 @@ const configureCompilerXmake = () => {
     /build:addon:[^\s]+/,
     'build:addon:xmake',
   );
+
+  packageJson.scripts['build:dev'] = 'npm run build:addon';
+  packageJson.scripts['prebuild:dev'] = 'xmake f -m debug';
   configureXmakeLua();
 };
 
@@ -232,12 +125,12 @@ const vscodeConfigureCCppPropertiesJson = () => {
       .map((item) => {
         delete item.configureFilter;
         delete item.configureSource;
-        item.includePath = libraryFolders();
+        item.includePath = libraryFolders(options);
         return item;
       });
     fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -246,9 +139,9 @@ const vscodeConfigureCompileFlagsTxt = () => {
   try {
     const filePath = path.join(__dirname, '..', 'compile_flags.txt');
     console.debug(`Configuring ${filePath} ...`.brightBlue);
-    fs.writeFileSync(filePath, twigCompile('compile_flags.txt', {folders: libraryFolders()}));
+    fs.writeFileSync(filePath, twigCompile('compile_flags.txt', {folders: libraryFolders(options)}));
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -266,7 +159,7 @@ const vscodeConfigureSettingsJson = () => {
     }
     fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -286,7 +179,7 @@ const configurePackageJson = () => {
     console.log(`Configuring ${filePath} ...`.brightBlue);
     fs.writeFileSync(path.join(__dirname, '..', 'package.json'), JSON.stringify(packageJson, null, 2));
   } catch (e) {
-    console.error(`Failed: ${e}`.red);
+    console.error(`Failed: ${e.message} on ${e.stack}`.red);
     process.exit(1);
   }
 };
@@ -310,7 +203,7 @@ program
 program.parse(process.argv);
 
 const options = program.opts();
-console.log(options);
+// console.log(options);
 
 /**
  * Configure default compiler
@@ -328,6 +221,9 @@ switch (options.buildSystem) {
     configureCompilerGyp();
 }
 
+/**
+ * Configure Visual Studio Code
+ */
 if (options.ide === supportedIdes.VSCODE) {
   vscodeConfigureCCppPropertiesJson();
   vscodeConfigureSettingsJson();
@@ -336,6 +232,9 @@ if (options.ide === supportedIdes.VSCODE) {
   }
 }
 
+/**
+ * Configure CLion
+ */
 if (options.ide === supportedIdes.CLION) {
   configureCMakeListsTxt();
 }
