+++ package.json
@@ -8,8 +8,7 @@
     "build:addon:cmake": "npx cmake-js configure && npx cmake-js build",
     "build:addon:gyp": "npx node-gyp configure && npx node-gyp build",
     "build:addon:xmake": "xmake",
-    "build:dev": "npm run build:addon",
-    "#build:dev": "npm run build:addon -- --debug",
+    "build:dev": "npm run build:addon -- --debug",
     "build": "npm run build:addon",
     "ca:lint": "npm run prettier:write && npm run lint:write",
     "ca:quality": "npm run jscpd && npm run depcruise",
@@ -29,7 +28,6 @@
     "lint:watch": "nodemon --exec 'npm run lint'",
     "lint:write": "npm run lint:cpp:write && npm run lint:js:write",
     "lint": "npm run lint:cpp && npm run lint:js",
-    "prebuild:dev": "npm run clean",
     "prebuild": "npm run clean",
     "prebuildc:dev": "npm run clean",
     "prebuildc": "npm run clean",
